// giveMeInsurace-client/src/app/shared/models/car.model.ts
export interface ICar {
    id: number;
    vehicleNumber: string;
    modelCode: string;
    year: number;
    road_ascent_date: string;
    description: string;
    air_bags: number;
    gear_type: number;
    esp: number;
    fcw: number;
    ldw: number;
    created_at: string;
    updated_at: string;
}